# config/swagger.yml
swagger: '2.0'
info:
  version: v1
  title: wecard
  description: The wecard APIs based on Swagger and built using Node.js
host: api.wecard.com
schemes:
  - http
  - https
basePath: /api
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
security:
  - Bearer: []
consumes:
  - application/json
  - text/json
produces:
  - 'application/json, text/json'
  - application/json
  - text/json
paths:
...
  /namecard:
    get:
      tags:
        - Namecard
      summary: Get all Namecards of current User
      operationId: namecard_getByCurrentUserId
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Namecard'
    post:
      tags:
        - Namecard
      summary: Create a Namecard
      operationId: namecard_create
      parameters:
        - $ref: '#/parameters/NamecardCreate'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Namecard'
    put:
      tags:
        - Namecard
      summary: Update the Namecard
      operationId: namecard_update
      parameters:
        - $ref: '#/parameters/NamecardUpdate'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Namecard'
...
definitions:
...
  Namecard:
    type: object
    properties:
      id:
        type: string
      cardType:
        type: string
      userId:
        type: string
      avatarUri:
        type: string
      name:
        type: string
      phone:
        type: string
      residence:
        type: string
      hometown:
        type: string
      company:
        type: string
      department:
        type: string
      occupation:
        type: string
      school:
        type: string
      major:
        type: string
      grade:
        type: integer
      exCompany:
        type: string
      exDepartment:
        type: string
      exOccupation:
        type: string
      nonpublic:
        type: boolean
      interests:
        type: array
        items:
          type: string
...
parameters:
...
  NamecardCreate:
    required: true
    name: body
    in: body
    description: The Namecard
    schema:
      allOf:
        - $ref: '#/definitions/Namecard'
      required:
        - cardType
        - name
        - phone
        - occupation
        - grade
  NamecardUpdate:
    name: body
    in: body
    required: true
    description: The Namecard
    schema:
      allOf:
        - $ref: '#/definitions/Namecard'
      required:
        - id
...